# backend/openapi.yaml
openapi: 3.0.3
info:
  title: HRMS API
  version: 0.1.0
  description: Sprint 0 契约占位；组织架构见 §6.2，完整清单见系分附录 H。
servers:
  - url: http://localhost:8080/api/v1
  - url: http://39.101.67.167:8080/api/v1
paths:
  /files:
    post:
      tags: [Common]
      summary: 上传本地文件（图片/文档）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 成功，返回 url / fileName / storedName
  /files/{storedName}:
    get:
      tags: [Common]
      summary: 下载或预览已上传文件
      security:
        - bearerAuth: []
      parameters:
        - name: storedName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 文件流
  /ai/capabilities:
    get:
      tags: [AI]
      summary: 当前角色可用快捷问题与可推送路由
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
  /ai/chat/stream:
    post:
      tags: [AI]
      summary: AI 流式对话（SSE）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                conversationId:
                  type: string
      responses:
        "200":
          description: text/event-stream；事件 data 为 JSON（type=delta|final|done|error）
  /ai/knowledge/docs:
    get:
      tags: [AI]
      summary: 知识库文档列表（仅 SYS_ADMIN）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
        "403":
          description: 非 SYS_ADMIN
    post:
      tags: [AI]
      summary: 上传知识库文档（仅 SYS_ADMIN）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
      responses:
        "200":
          description: 成功，返回文档元数据
  /ai/knowledge/docs/{id}/enabled:
    put:
      tags: [AI]
      summary: 启停知识库文档（仅 SYS_ADMIN）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: 成功
  /ai/knowledge/docs/{id}:
    delete:
      tags: [AI]
      summary: 删除知识库文档及向量（仅 SYS_ADMIN）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
  /approvals/handover-candidates:
    get:
      tags: [Approval]
      summary: 正式离职交接人选人（轻量搜索，非花名册）
      description: |
        供审批中心确认工作交接时搜索在职员工。
        允许 HR_STAFF / DEPT_MANAGER / FINANCE_MANAGER / SYS_ADMIN；
        不开放给普通 FINANCE 专员；不经过花名册 DataScope。
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 姓名或工号关键词
      responses:
        "200":
          description: 成功；data 为候选列表（employeeId/name/empNo/department）
        "403":
          description: 无权限
  /approvals/delegate-candidates:
    get:
      tags: [Approval]
      summary: 审批委托候选人（仅有审批权限的启用用户）
      description: |
        供「审批委托」选人：仅返回具备审批能力的启用账号
        （角色 SYS_ADMIN / HR_STAFF / DEPT_MANAGER / FINANCE_MANAGER，
        或权限 approval:handle / approval:action / menu:workflow / menu:approval），
        并排除当前登录用户。
        keyword 为空时返回前若干条；可按姓名、工号、用户名模糊搜索。
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 姓名 / 工号 / 用户名关键词
      responses:
        "200":
          description: 成功；data 为候选列表（userId/name/username/empNo/department）
  /workbench/summary:
    get:
      tags: [Workbench]
      summary: 工作台 KPI 汇总；pendingApprovals 为当前登录用户待办数（与审批中心待办同口径）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功；单项统计失败时降级为 0/空列表
  /system/users:
    get:
      tags: [System]
      summary: 用户分页
      security:
        - bearerAuth: []
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 成功
    post:
      tags: [System]
      summary: 新增用户
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功，data.userId
  /system/users/{id}:
    put:
      tags: [System]
      summary: 更新用户状态/角色
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
  /system/roles:
    get:
      tags: [System]
      summary: 角色列表（预置，不可新增删除）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
  /system/roles/{id}:
    put:
      tags: [System]
      summary: 编辑角色名称（编码只读）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
  /system/roles/{id}/permissions:
    put:
      tags: [System]
      summary: 分配角色权限（变更后失效权限缓存）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
  /system/permissions:
    get:
      tags: [System]
      summary: 权限码列表
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
  /system/login-logs:
    get:
      tags: [System]
      summary: 登录日志分页
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
  /system/operation-logs:
    get:
      tags: [System]
      summary: 操作审计日志分页
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
  /departments/tree:
    get:
      tags: [Org]
      summary: 部门树（含人数、负责人）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
  /departments/{id}/headcount:
    get:
      tags: [Org]
      summary: 部门人数（本部门 + 含下属）
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeptId"
      responses:
        "200":
          description: 成功
  /departments/{id}/can-delete:
    get:
      tags: [Org]
      summary: 删除前校验
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeptId"
      responses:
        "200":
          description: 成功
  /departments:
    post:
      tags: [Org]
      summary: 新增部门
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeptRequest"
      responses:
        "200":
          description: 成功，data.id
  /departments/{id}:
    get:
      tags: [Org]
      summary: 部门详情（树节点同构字段，无 children）
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeptId"
      responses:
        "200":
          description: 成功
    put:
      tags: [Org]
      summary: 编辑部门（可移动上级）
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeptId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeptRequest"
      responses:
        "200":
          description: 成功
    delete:
      tags: [Org]
      summary: 删除部门（须无子部门、无归属员工、无职位）
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeptId"
      responses:
        "200":
          description: 成功
        "422":
          description: 30005 部门非空
  /departments/{id}/merge:
    put:
      tags: [Org]
      summary: 部门合并（自动转移员工与子部门；含已离职，避免源部门删除后悬空）
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeptId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetDepartmentId]
              properties:
                targetDepartmentId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: 成功
  /positions:
    get:
      tags: [Org]
      summary: 职位分页列表（activeCount=归属人数：试用/正式/待离职，不含已离职）
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: departmentId
          in: query
          schema:
            type: integer
            format: int64
        - name: sequenceCode
          in: query
          schema:
            type: string
            enum: [M, P, S]
      responses:
        "200":
          description: 成功；list[].activeCount 为归属人数（含待离职 30）
    post:
      tags: [Org]
      summary: 新增职位
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePositionRequest"
      responses:
        "200":
          description: 成功，data.id
  /positions/{id}:
    get:
      tags: [Org]
      summary: 职位详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
      responses:
        "200":
          description: 成功
    put:
      tags: [Org]
      summary: 编辑职位
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePositionRequest"
      responses:
        "200":
          description: 成功
    delete:
      tags: [Org]
      summary: 删除职位
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PositionId"
      responses:
        "200":
          description: 成功
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeptId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    PositionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    CreateDeptRequest:
      type: object
      required: [name, deptCode, sortOrder]
      properties:
        name:
          type: string
          maxLength: 64
        deptCode:
          type: string
          pattern: "^[A-Za-z0-9]{2}$"
        parentId:
          type: integer
          format: int64
          nullable: true
        headEmployeeId:
          type: integer
          format: int64
          nullable: true
        sortOrder:
          type: integer
          description: >-
            同级（同一 parentId）排序序号。新建须严格大于同级已有最大序号（建议默认 max+1）；
            编辑若变更上级或序号，亦须大于新同级中其他部门的最大序号。
        description:
          type: string
          maxLength: 256
    UpdateDeptRequest:
      allOf:
        - $ref: "#/components/schemas/CreateDeptRequest"
    CreatePositionRequest:
      type: object
      required:
        - name
        - sequenceCode
        - gradeMin
        - gradeMax
        - defaultProbationMonths
        - isStandard
      properties:
        name:
          type: string
        sequenceCode:
          type: string
          enum: [M, P, S]
        departmentId:
          type: integer
          format: int64
          nullable: true
        gradeMin:
          type: string
        gradeMax:
          type: string
        defaultProbationMonths:
          type: integer
        isStandard:
          type: boolean
        description:
          type: string